• Moved SampleSpec and SampleIterator to globals/sampleiterator
