+++ psyneulink/core/globals/sampleiterator.py
@@ -1,0 +1,357 @@
+#
+# Princeton University licenses this file to You under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.  You may obtain a copy of the License at:
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed
+# on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
+#
+#
+# *******************************************   SAMPLER CLASSES ********************************************************
+'''
+
+* `SampleSpec`
+* `SampleIterator`
+
+'''
+
+import numpy as np
+
+import typecheck as tc
+from collections import Iterator
+from decimal import Decimal, getcontext
+
+
+from psyneulink.core.components.functions.function import is_function_type
+
+
+__all__ = ['SampleSpec', 'SampleIterator']
+
+
+# Number of decimal places used to set precision of decimal module
+SAMPLE_SPEC_PRECISION = 16
+
+
+class SampleGeneratorError(Exception):
+    def __init__(self, error_value):
+        self.error_value = error_value
+
+
+class SampleSpec():
+    '''
+    SampleSpec(    \
+    start=None,    \
+    stop=None,     \
+    step=None,     \
+    num=None,      \
+    function=None  \
+    precision=None \
+    )
+
+    Specify the information needed to create a SampleIterator that will either (a) generate discrete values in a range
+    or (b) call a function can generate continous values within the range.
+
+    (1) Generate values in a range by explicitly specifying a finite regular sequence of values, using an appropriate
+        combination of the **start**, **stop**, **step** and/or **num**arguments.
+
+        * if **start**, **stop**, and **step** are specified, the behavior is similar to `Numpy's arange
+          <https://docs.scipy.org/doc/numpy-1.13.0/reference/generated/numpy.arange.html>`_. Calling
+          `next <SampleIterator.__next__>`first returns **start**. Each subsequent call to next returns
+          **start** :math:`+` **step** :math:`*` current_step.
+          Iteration stops when the current value exceeds the **stop** value.  If any of the specified arguments are
+          floats, precision determines the number of decimal places used for rounding to ensure num is a int.
+
+        * if **start**, **stop**, and **num** are specified, the behavior is similar to `Numpy's linspace
+          <https://docs.scipy.org/doc/numpy-1.13.0/reference/generated/numpy.linspace.html>`_. Calling
+          `next <SampleIterator.__next__>` first returns **start**. Each subsequent call to next returns
+          **start** :math:`+` step :math:`*` current_step, where step is set to :math:`\\frac{stop-start)}{num - 1}`.
+          Iteration stops when **num** is reached, or the current value exceeds the **stop** value.
+
+        * if **start**, **stop*, **step**, and **num** are all specified, then **step** and **num** must be compatible.
+
+    (2) Specify a function, that is called repeatedly to generate a sequence of values.
+
+        * if **num** is specified, the **function** is called once on each iteration until **num** iterations are
+          complete.
+
+        * if **num** is not specified, the **function** is called once on each iteration, and iteration may continue
+          indefintely.
+
+    .. note::
+        * Some OptimizationFunctions may require that their SampleIterators have a "num" attribute.
+        * The Python decimal module is used to implement **precision** for rounding.
+
+
+    Arguments
+    ---------
+
+    start : int or float
+        First sample in the sequence
+
+    stop : int or float
+        Maximum value of last sample in the sequence
+
+    step : int or float
+        Space between samples in sequence
+
+    num :  int
+        Number of samples
+
+    function : function or function class
+        Function to be called on each iteration. Must return one sample each time it is called.
+
+    precision : int default 16
+        Number of decimal places used for rounding in floating point operations.
+
+
+    Attributes
+    ----------
+
+    start : float
+        First sample in the sequence
+
+    stop : float
+        Maximum value of last sample in the sequence
+
+    step : float
+        Space between samples in sequence
+
+    num :  int
+        Number of samples
+
+    function : function
+        Function to be called on each iteration. Must return one sample.
+
+    '''
+    @tc.typecheck
+    def __init__(self,
+                 start:tc.optional(tc.any(int, float))=None,
+                 stop:tc.optional(tc.any(int, float))=None,
+                 step:tc.optional(tc.any(int, float))=None,
+                 num:tc.optional(int)=None,
+                 function:tc.optional(is_function_type)=None,
+                 precision:tc.optional(int)=None
+                 ):
+
+        self._precision = precision or SAMPLE_SPEC_PRECISION
+        # Save global precision for later restoration
+        _global_precision = getcontext().prec
+        # Set SampleSpec precision
+        getcontext().prec = self._precision
+
+        if function is None:
+            if start is None or stop is None:
+                raise SampleGeneratorError("If 'function' is not specified, then 'start' and 'stop' must be "
+                                                "specified.")
+            if num is None and step is not None:
+                num = int(Decimal(1.0) + Decimal(stop - start) / Decimal(step))
+                # num = 1.0 + (stop - start) / step
+            elif step is None and num is not None:
+                step = (Decimal(stop) - Decimal(start)) / (num - 1)
+            elif num is None and step is None:
+                raise SampleGeneratorError("Must specify one of {}, {} or {}."
+                                                .format(repr('step'), repr('num'), repr('function')))
+            else:
+                if not np.isclose(num, 1.0 + (stop - start) / step):
+                    raise SampleGeneratorError("The {} ({}) and {} ({}} values specified are not comaptible."
+                                                    .format(repr('step'), step, repr('num'), num))
+
+        elif is_function_type(function):
+            if start is not None:
+                raise SampleGeneratorError("Only one of {} ({}) and {} ({}} may be specified."
+                                                .format(repr('start'), start, repr('function'), function))
+            if step is not None:
+                raise SampleGeneratorError("Only one of {} ({}) and {} ({}} may be specified."
+                                                .format(repr('step'), step, repr('function'), function))
+        else:
+            raise SampleGeneratorError("{} is not a valid SampleSpec function."
+                                            .format(function))
+
+        # FIX: ELIMINATE WHEN UPGRADING TO PYTHON 3.5.2 OR 3.6, (AND USING ONE OF THE TYPE VERSIONS COMMENTED OUT ABOVE)
+        # Validate entries of specification
+        #
+        self.start = start
+        self.stop = stop
+        self.step = step
+        self.num = num
+        self.function = function
+
+        # Restore global precision
+        getcontext().prec = _global_precision
+
+
+class SampleIterator(Iterator):
+    """
+    SampleIterator(               \
+    specification                 \
+    )
+
+    Creates an iterator that returns the next sample from a sequence on each call to `next <SampleIterator.__next__>`.
+
+    The pattern of the sequence depends on the **specification**, which may be a list, nparray, or SampleSpec. Most of
+    the patterns depend on the "current_step," which is incremented on each iteration, and set to zero when the
+    iterator is reset.
+
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | **Specification**              |  **what happens on each iteration**       | **StopIteration condition**        |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | list, nparray                  | look up the item with index current_step  | list/array ends                    |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | SampleSpec(start, stop, step)  | start + step*current_step                 | current_step = num or value > stop |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | SampleSpec(start, stop, num)   | start + step*current_step                 | current_step = num or value > stop |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | SampleSpec(function, num)      | call function                             | current_step = num                 |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+    | SampleSpec(function)           | call function                             | iteration does not stop            |
+    +--------------------------------+-------------------------------------------+------------------------------------+
+
+    .. note::
+        We recommend reserving the list/nparray option for cases in which the samples do not have a pattern that can be
+        represented by a SampleSpec, or the number of samples is small. The list/nparray option requires all of the
+        samples to be stored and looked up, while the SampleSpec options generate samples as needed.
+
+    """
+
+    @tc.typecheck
+    def __init__(self,
+                 specification:tc.any(list, np.ndarray, range, SampleSpec)):
+        '''
+
+        Arguments
+        ---------
+
+        specification : list or SampleSpec
+            specifies what to use for `generate_current_value <SampleIterator.generate_current_value>`.
+
+        Attributes
+        ----------
+
+        start : scalar or None
+            first sample in the sequence, or None
+
+        stop : scalar or None
+            last sample in the sequence, or None
+
+        step : scalar
+            space between samples in the sequence, or None
+
+        num : int or None
+            number of values returned before the iterator stops. If None, the iterator may be called indefinitely.
+
+        current_step : int
+            number of current iteration
+
+        Returns
+        -------
+
+        List(self) : list
+        '''
+
+        # FIX: DEAL WITH head?? OR SIMPLY USE CURRENT_STEP?
+        # FIX Are nparrays allowed? Below assumes one list dimension. How to handle nested arrays/lists?
+
+        if isinstance(specification, range):
+            specification = list(specification)
+
+        if isinstance(specification, list):
+            self.start = specification[0]
+            self.stop = None
+            self.step = None
+            self.num = len(specification)
+            self.generator = specification                       # the list
+
+            def generate_current_value():                        # index into the list
+                return self.generator[self.current_step]
+
+        elif isinstance(specification, SampleSpec):
+
+            if specification.function is None:
+                self.start = specification.start
+                self.stop = specification.stop
+                # self.step = Fraction(specification.step)
+                self.step = specification.step
+                self.num = specification.num
+                self.generator = None                    # ??
+
+                def generate_current_value():   # return next value in range
+                    from decimal import Decimal, getcontext
+                    # Save global precision for later restoration
+                    _global_precision = getcontext().prec
+                    # Set SampleSpec precision
+                    getcontext().prec = specification._precision
+                    return_value = float(Decimal(self.start) + Decimal(self.step) * Decimal(self.current_step))
+                    # Restore global precision
+                    getcontext().prec = _global_precision
+                    return return_value
+
+            elif is_function_type(specification.function):
+                self.start = 0
+                self.stop = None
+                self.step = 1
+                self.current_step = 0
+                self.num = specification.num
+                self.head = self.start
+                self.generator = specification.function
+
+                def generate_current_value():  # call function
+                    return self.generator()
+
+            else:
+                assert False, 'PROGRAM ERROR: {} item of {} passed to specification arg of {} ' \
+                              'is not an iterator or a function_type'.\
+                              format(repr('function'), SampleSpec.__name__, self.__class__.__name__)
+
+
+        else:
+            assert False, 'PROGRAM ERROR: {} argument of {} must be a list or {}'.\
+                          format(repr('specification'), self.__class__.__name__, SampleSpec.__name__)
+
+        self.current_step = 0
+        self.head = self.start
+        self.generate_current_value = generate_current_value
+
+    def __next__(self):
+        """
+
+        :return:
+        Sample value for the current iteration.
+        """
+        if self.num is None:
+            current_value = self.generate_current_value()
+            if hasattr(self, 'stop'):
+                if self.stop is not None:
+                    if current_value <= self.stop:
+                        return current_value
+                    else:
+                        raise StopIteration
+            return current_value
+        if self.current_step < self.num:
+            current_value = self.generate_current_value()
+            self.current_step += 1
+            if hasattr(self, 'stop'):
+                if self.stop is not None:
+                    if current_value <= self.stop:
+                        return current_value
+                    else:
+                        raise StopIteration
+            return current_value
+
+        else:
+            raise StopIteration
+
+    def __iter__(self):
+        self.current_step = 0
+        return self
+
+    def __call__(self):
+        return list(self)
+
+    def reset(self, head=None):
+        '''Reset iterator to a specified item
+        If called with None, resets to first item (if `generator <SampleIterators.generator>` is a list or
+        deterministic function.
+        '''
+
+        self.current_step = 0
+        self.head = head or self.start
